refactor: judgement 패키지를 만들어 볼 판정에 관한 클래스 리팩키징
